+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 1.1.1 (not yet released)
 
+- changed default boot timeout: 20 seconds for standalone server
+  and 2 minutes for managed domain
+- added `CliScript`
+- improved waiting for server: `OnlineOptions.connectionTimeout`
+  is now also used when connecting to the server fails
+
 ## 1.1.0 [The Wintermute Edition](https://en.wikipedia.org/wiki/Neuromancer) (2016-04-25)
 
 - added `OnlineOptions.bootTimeout`